feat(Application) introduce Mapperly mappings for NotificationMessage<EmailExpectingPayment>
